♻️ refactor: 크레딧 이미지 경로 수정
